Refactor ops into a submodule
This is first part of refactoring, with the following steps:
* ops.py was moved into ops/__init__.py
* Utilities related to keeping the names of registered operators were
  moved to a dedicated _registry.py file.
* Utilities for parsing and reporting name based on schema name were
  moved to a dedicated _names.py file.
* Docstring generation was moved to dedicated _docs.py file.
* Some docstrings were added.

The visibility of the symbols placed in the new files was kept the same
as in the original code to not cause cascading changes in DALI.

The next steps will involve:
* Extending the registration of the custom operator to allow
  moving the code into separate files (like PythonOp or TFRecordReader).
* Adjusting the processing of inputs and arguments, and filling the
  OpSpec, so the code can better maintained between the two APIs.
